+++ template/hono/src/resources/src/pages/auth/resetPassword.vue
@@ -40,7 +40,7 @@
           <Button
             type="submit"
             label="Reset Password"
-            class="btn btn-primary d-grid w-100"
+            class="btn btn-primary w-100"
             icon="bi bi-key ms-2"
             :disabled="processing" />
         </form>